Commit 1: Renamed EncounterGenerator.py to encounter_generator.py to be consistent with other generator names.
