+++ thedmstoolkit/urls.py
@@ -21,7 +21,7 @@ from django.urls import path
 from toolkit.views.character_generator.character_generator import CharacterGenerator
 from toolkit.views.confirm_account_creation import ConfirmAccountCreation
 from toolkit.views.create_account import CreateAccount
-from toolkit.views.encounter_generator.EncounterGenerator import EncounterGenerator
+from toolkit.views.encounter_generator.encounter_generator import EncounterGenerator
 from toolkit.views.home_page import HomePage
 from toolkit.views.login import Login
 from toolkit.views.logout import Logout
